network packet analzyer/requirements_gui.txt
scapy>=2.4.5
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
Pillow>=8.3.0 